docs: add Podman machine reboot step to cert install (#13498)
This adds an optional step to the end of the instructions for installing
additional root certificates to a Podman machine. On MacOS, with the podman
machine default as my container host, "podman compose up" would continue to
fail with self-signed certificate errors after following the instructions in
this guide, until I fully rebooted the Podman machine.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>
Co-authored-by: Shipra Singh <94683525+shipsing@users.noreply.github.com>