feat(managed-by): apply default settings on init (#15089)
* feat(managed-by): apply default settings on init

### What does this PR do?

When a setting exists in your `default-settings.json`, but does not
exist in your user `settings.json`. The setting will then be set in your `settings.json` if it does not exist.

This "applies" the default-settings.json to your settings.json on
first boot to any values that aren't there yet.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A, no UI change

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/podman-desktop/podman-desktop/issues/15087

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

1. Add the following setting to your default-settings.json configuration
   (different on Linux and Windows, see: https://podman-desktop.io/docs/configuration/managed-configuration):

```
cat /Library/Application\ Support/io.podman_desktop.PodmanDesktop/default-settings.json
{
  "telemetry.enabled": false
}
```

2. Check that the value does NOT exist yet in your local
   `settings.json`:

```
cat ~/.local/share/containers/podman-desktop/configuration/settings.json
{
  "window.bounds": {
    "x": 1959,
    "y": 242,
    "width": 1255,
    "height": 1064
  }
}
```

3. Boot up Podman Desktop (with this PR)

4. Check the console logs have: `[Managed-by]: Applied default settings
   for: ...`

5. Check that the value was added to settings.json:

```
cat ~/.local/share/containers/podman-desktop/configuration/settings.json
{
  "window.bounds": {
    "x": 1959,
    "y": 242,
    "width": 1255,
    "height": 1064
  },
  "telemetry.enabled": false
}
```

On subsequent startups, the value won't be overriden (it checks if the
value exists before doing anything to `settings.json`).

You can also check the reverse, have the `telemetry.enabled` setting
already in your `settings.json` and check to see if the
`default-settings.json` had overridden the setting.

- [X] Tests are covering the bug fix or the new feature

Signed-off-by: Charlie Drage <charlie@charliedrage.com>

* chore: add more tests, make sure we saveDefault if changes.

Adds more tests to cover the use-case of making sure that if a value is
already default in the schema, it does NOT copy over to setings.json (no
point to have it there... as it's already set by default).

We also make sure that we call saveDefault if there are any
configuration changes. In the previous commit, it would only save if
the user explicitly changes it (saveDefault is then called, and then
checks the schema and saves as necessary).

Signed-off-by: Charlie Drage <charlie@charliedrage.com>

---------

Signed-off-by: Charlie Drage <charlie@charliedrage.com>